continue_to validation
